chore: refactor bindings and entrypoints
