+++ packages/cli/tests/bindings-test/src/test_assets.py
@@ -1,5 +1,5 @@
 import pytest
-from workers._workers import _FetcherWrapper
+from workers.rpc import _FetcherWrapper
 
 
 @pytest.mark.asyncio
